+++ proviscore/cursillos.go
@@ -1,14 +1,14 @@
-package provisCore
+package proviscore
 
 import (
 	"context"
-	"github.com/angelbarreiros/ProvisGo/provisEntities"
+	"github.com/angelbarreiros/ProvisGo/provisentities"
 	"github.com/angelbarreiros/ProvisGo/util"
 	"net/http"
 	"net/url"
 )
 
-func (pe provisExecutor) Cursillos(filterParams *provisEntities.CursillosParams) (*provisEntities.CursillosResponse, *provisEntities.ErrorResponse) {
+func (pe provisExecutor) Cursillos(filterParams *provisentities.CursillosParams) (*provisentities.CursillosResponse, *provisentities.ErrorResponse) {
 	ctxWithTimeout, cancel := context.WithTimeout(context.Background(), pe.defaultTimeout)
 	defer cancel()
 	resultChan := make(chan util.RequestResult, 1)
@@ -25,7 +25,7 @@ func (pe provisExecutor) Cursillos(filterParams *provisEntities.CursillosParams)
 			nil)
 		request = request.WithContext(ctxWithTimeout)
 
-		var responseArray []provisEntities.Cursillo = make([]provisEntities.Cursillo, 0)
+		var responseArray []provisentities.Cursillo = make([]provisentities.Cursillo, 0)
 		var responseBody = &responseArray
 		result := util.ExecuteRequest(ctxWithTimeout, pe.client, request, pe.config.Debug, responseBody)
 		resultChan <- result
@@ -34,12 +34,12 @@ func (pe provisExecutor) Cursillos(filterParams *provisEntities.CursillosParams)
 	select {
 	case res := <-resultChan:
 		if res.Response != nil {
-			var responseCursillos = res.Response.(*[]provisEntities.Cursillo)
-			return &provisEntities.CursillosResponse{Cursillos: *responseCursillos}, res.Error
+			var responseCursillos = res.Response.(*[]provisentities.Cursillo)
+			return &provisentities.CursillosResponse{Cursillos: *responseCursillos}, res.Error
 		}
 		return nil, res.Error
 	case <-ctxWithTimeout.Done():
-		return nil, &provisEntities.ErrorResponse{
+		return nil, &provisentities.ErrorResponse{
 			Code:    http.StatusRequestTimeout,
 			Message: "Request timeout: operation cancelled after 10 seconds",
 		}